# src/doc/swagger_doc.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My Book List RestAPI",
    "description": "Livros e Listas para todos os gostos!"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/friendship/searchUserFriendship": {
      "get": {
        "tags": [
          "friendship"
        ],
        "description": "Buscar de amizades do usuário.",
        "parameters": [
          {
            "name": "idUser",
            "description": "Id do usuário a buscar as amizades.",
            "required": true,
            "type": "integer",
            "in": "query"
          },
          {
            "name": "page",
            "description": "Página de itens.",
            "required": true,
            "type": "integer",
            "in": "query"
          },
          {
            "name": "itens",
            "description": "Quantidade de itens a buscar.",
            "required": true,
            "type": "integer",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "object",
                  "properties": {
                    "friendshipList": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "idUser1": {
                            "type": "number",
                            "example": 0
                          },
                          "idUser2": {
                            "type": "number",
                            "example": 0
                          },
                          "dateOfFriendship": {
                            "type": "string",
                            "example": "yyyy-MM-dd 00:00:00.000"
                          },
                          "friend": {
                            "type": "object",
                            "properties": {
                              "idUser": {
                                "type": "number",
                                "example": 0
                              },
                              "name": {
                                "type": "string",
                                "example": "name"
                              },
                              "nickname": {
                                "type": "string",
                                "example": "nickname"
                              },
                              "photo": {
                                "type": "string",
                                "example": "base64 or null"
                              }
                            }
                          }
                        }
                      }
                    },
                    "pageCount": {
                      "type": "number",
                      "example": 1
                    }
                  }
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/friendship/verifyFriendship": {
      "get": {
        "tags": [
          "friendship"
        ],
        "description": "Verificar a amizade entre dois usuários.",
        "parameters": [
          {
            "name": "idUser1",
            "description": "Id do usuário a buscar a amizade.",
            "required": true,
            "type": "integer",
            "in": "query"
          },
          {
            "name": "idUser2",
            "description": "Id do usuário a buscar a amizade.",
            "required": true,
            "type": "integer",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "example": "true or false"
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/friendship/createFriendship": {
      "post": {
        "tags": [
          "friendship"
        ],
        "description": "Criar amizade entre dois usuários.",
        "parameters": [
          {
            "name": "create",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "idUser1": {
                  "type": "number",
                  "example": 0
                },
                "idUser2": {
                  "type": "number",
                  "example": 0
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "example": "true or false"
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/friendship/deleteFriendship": {
      "delete": {
        "tags": [
          "friendship"
        ],
        "description": "Deletar a amizade entre dois usuários.",
        "parameters": [
          {
            "name": "idUser1",
            "description": "Id do usuário a deletar a amizade.",
            "required": true,
            "type": "integer",
            "in": "query"
          },
          {
            "name": "idUser2",
            "description": "Id do usuário a deletar a amizade.",
            "required": true,
            "type": "integer",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "string",
                  "example": "true or false"
                },
                "status": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro.",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Mensagem de erro."
                },
                "status": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}